Activity 9, Accordion updates
Usage, tab keys get to accordian panel buttons and the open panel. Open and close panels when on accordian button and press space or enter key. Use up and down arrows for currunt accordian  on only to go from accordion button of that current accordion.
Update: Accordian headers and keyboard focusabel, Accordion headers are announced as buttons instead of list items. Accordion headers open panels with a click or key press, Accordion announce expanded when a panel is opened and collapsed when closed. Accordion panels are focusable with a Tab key press when opened. Navigation between accordion headers with up and down arrow keys and tab key.