Activity 10, Tab Panel Updates
List of semantics are replaced with tab panel semantics. Focus tab position in the tablist is announced. When a tabl is in focus, its associated panel displays. Arrow keys can be used to navigate between tabs. Tab key can be used to move from a selected tab directly to its associated panel Shift + Tab to move back to tabs. Panels are keyboard focusable.