Activity 13 Update of Tree Menu
Tree menu receives focus and instuction are announced on how to use menu with a keyboard. When navigating through the tree with a keybaord, elements are announced with tree menu semantics. When a tree mneu item with a submenu receive focus, the submenu state is announed as expanded when open or collapsed when closed. Only elements of the tree menu that are in view are able to receive focus. Tree menu functions with a keyboard. Tab key navigate into tree menu, first tree item,and exit tree menu. Up and down arrows are used to move between tree items. The space or Enter key are used to expand/collapse a tree item with a submenu. When submenu is opened, focus moves to the first tree item in menu.