Activity 14 Update Sortable LIst
Instructions are announced on using the sortable list with a keyboard when it first receives focus. When navigating through list items, their psoition is announced along with an indication they can be moved. Keyboard operation, Tab key is used to navigate into the list, and exit the list. Up/Down arrows are used to move between list items. Window Ctrl (mac Command) + Up/Down arrow key selects a list item and moves it to a new location.